api/requirements.txt
Django==1.11
gunicorn==19.6.0
djangorestframework
